generator: Minor refactoring
Move some code around with no changes to the bindings.
